Add PR approval check for specific directories (#31)
Adds process/CI workflow to check that the approvers for every PR are in
a committee of recognized approvers. This is a much simplified version
of the r?bot that rust-lang uses.

## Testing Scenarios

### Happy path

What happens when 2 of the approvers are in the committee

Run:
https://github.com/jaisnan/rust-dev/pull/16/checks?check_run_id=26914353663

## What if someone not in the list approves, and 1 from the committee
approves?

In this scenario, we have a committee that consists of someone called
"jaisu-1". But if the approvals came from one ID that's recognized, and
another called "Jaisu-1" (Note to the reader: "jaisu-1" != "Jaisu-1"),
then the workflow fails and the PR merge is (rightfully) blocked.

Run:
https://github.com/jaisnan/rust-dev/pull/15/checks?check_run_id=26914179444

## Call-Outs

1. We need to add a requirement through settings that at least 2
approvers are required before merging, and allow anyone to approve (if
it's not already the setting).
2. Once the first iteration of a committee is finalized, we can merge
this workflow to begin the approval checking.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.