`kani list` Workflow (#146)
Adds a workflow that runs `kani list` on the standard library and posts
the results in a comment on the pull request. This workflow runs iff a
pull request is merged, so that we have one comment at the end of a PR
with the most up-to-date list results. (I considered other approaches,
like running it on opening the PR or with every commit, but decided
having one single comment with the final changes was best for
brevity/clarity).

See this [test
PR](https://github.com/carolynzech/verify-rust-std/pull/10) on my fork
as a demo of how it would work.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Celina G. Val <celinval@amazon.com>