Cleanup stray comment (#218)
With https://github.com/model-checking/kani/issues/3670 having been
addressed we re-enabled harnesses in #211. Yet the comment explaining
previously-commented-out-harnesses stayed in place.

Removing it for it no longer applies.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.