Improve `u16::carrying_mul` harness performance (#230)
This harness takes between 22-25 minutes in CI. Change the harnesses's
macro to use kissat solver instead, which brings verification time down
to ~7 seconds.

As an aside, I noticed this problem because the partition 1 runner in
#229 is taking ~55 minutes, where the other partitions are taking ~30
minutes. This harness accounts for almost the entire difference. One
nice consequence of partitioning verification is that it will make
performance issues like this more noticeable.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.