Run format check in our CI and fix repo format (#205)
Changed the `check-rustc.sh` script to run the format check. Also added
a `--bless` mode to apply the changes locally.
This change also includes the result of running the script with the
`--bless` mode on.

Related to #38 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>
Co-authored-by: Carolyn Zech <cmzech@amazon.com>