Subtree update automation: remove extraneous library folder (#288)
This restores a fix from #250 that was deemed no longer necessary with
the move to subtree split (in #270), but this proved to be wrong. (See
https://github.com/model-checking/verify-rust-std/pull/275/commits/ab648fe7196e1016d22fac18bc80eabb17e6e5a8
for such an example of an unintended change.)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

Co-authored-by: thanhnguyen-aws <ntson@amazon.com>