Track metrics for `core` and `std` crates, add loop metrics (#258)
The crate that we compute metrics for is now configurable when invoking
`kani_std_analysis.py`. Should we find a need for further crates beyond
`core` or `std` we just need to invoke it more times.

We are now also reporting loop/no-loop information across all the kinds
of functions we are tracking as metrics.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.