Subtree update automation: do not use `git subtree merge --squash` (#292)
`git subtree merge --squash` will always reset the subtree to the state
of the tree that is being merged from, which effectively eradicates all
our local changes. This was just masked by merge conflicts arising as we
were always attempted to merge from some long-ago version as we hadn't
consistently kept "git-subtree-split" markers.

This PR now amends the pull request descriptions to make sure we retain
the necessary information and uses `git merge` instead of the subtree
command.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.