Sync the VeriFast proofs and provide guidance on same (#313)
Updates the VeriFast proofs after `linked_list.rs` was modified by
https://github.com/rust-lang/rust/commit/c39f33baae94665d69a45d45a00d0dc028c80cc9
.

Also:
- Added a bash script that attempts to automatically patch the proofs
after the original file was changed.
- The VeriFast CI actions now produce an error alert suggesting to run
this script, if a source file that is the subject of a VeriFast proof is
changed.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.