Subtree update automation: create merge PR even when subtree/library is up-to-date (#326)
When the PR updating the `subtree/library` branch has been merged
already we may still need to bring updates into `main` that haven't
previously been added to a pull request. This situation arises when an
older merge-into-main PR was still open at the time where the subtree
automation creating the `subtree/library` update PR ran.

The changes in this PR avoid this problem by making sure the
merge-into-main part of the automation is run even when
`subtree/library` is already up-to-date.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.