Only require two approvals for changes to `doc/`, `library/` or `verifast-proofs/` (#351)
We are making a lot of changes to our `scripts/` and
`.github/workflows/` folders to improve automation. These changes are
low-risk; we have no plans to upstream them and they're not part of the
public-facing challenge effort. Update our workflow to require two
committee approvals iff `doc/`, `library/` or `verifast-proofs/` are
modified, and otherwise just require one.

See https://github.com/carolynzech/rust/pull/35 and
https://github.com/carolynzech/rust/pull/36 for test runs on my fork.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.