Fix num::nonzero::NonZero::<*>::rotate_{left,right} contracts (#346)
The result of a rotate operation is always non-zero, which could still
be less than zero in case of signed types.

Proofs were failing with the prior contract, but this still passed CI as
we haven't yet picked up the Kani version where autoharness exits with a
non-zero exit code in case of proof failure.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.