Challenge 7: make panicking part optional; remove mention of verifying intrinsics (#357)
Update the atomic intrinsics challenge to make verifying the intrinsics
out of scope, since that would require verifying compiler
implementations that are out of scope for this effort. Also make the
part about verifying the absence of panics optional to restrict the
requirements to safety only.

I also added a `mdbook-linkcheck` dependency because I discovered I
needed that to build the book locally.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.