additional transmute and transmute_unchecked harnesses (#264)
Toward solving #19 

This pr adds some additional harnesses for transmute_unchecked and
transmute.

With this, we suspect that the main harnesses for part 1 of challenge 1
(verifying the transmute intrinsics directly) are there (besides adding
more of the same kinds of harnesses, like the 2-way harnesses for more
types). We go into more detail about what needs to be done for part 1,
as well as what has been done and what can't be done here:
https://docs.google.com/document/d/1zFGANNMx8mZ8fucKrN--ELwKASUPeP20THH6M_fQ7jg/edit?usp=sharing

Just one note: transmute has far fewer harnesses than
transmute_unchecked here -- this is because it is not currently possible
to write a wrapper for transmute, and it is thus not possible to write a
function contract. A lot of the harnesses for transmute_unchecked test
the function contract's validity clause, rather than the function
itself, explaining this disparity.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: AlexLB99 <marcgolvanian@gmail.com>
Co-authored-by: Michael Tautschnig <tautschn@amazon.com>