Use uname -m instead of uname -p (#372)
On my Debian machine, uname -p returns "unknown" and the manpage says -p
is non-portable. I do get x86_64 on from uname -m and that doesn't have
the non-portable disclaimer on it.

> Please add a description of your PR.
> If this is a solution to an open challenge, please explain your
solution.
>
> Don't forget to check our book to ensure your solution satisfy the
overall
> requirements as well as the challenge success criteria.
>

Resolves #ISSUE-NUMBER

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.