update-subtree action: patch VeriFast proofs (#371)
When preparing a merge subtree PR, try to patch the VeriFast proofs.

Also tweaks `patch-verifast-proofs.sh` to ensure that no temporary files
are left behind inside the working directory, even when the script fails
mid-way.

Note: I did not test this.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Michael Tautschnig <mt@debian.org>