Subtree update automation: update submodules (#394)
Update submodules to the exact version that upstream uses at the commit
that we are trying to merge towards. We never modify the contents of
submodules, so we don't need to make any attempt to merge changes and
instead can directly pick the right upstream version.

See
https://github.com/tautschnig/verify-rust-std/actions/runs/15846078305/job/44668465448
for a successful run and
https://github.com/tautschnig/verify-rust-std/pull/25 for the ensuing PR
(which, notably, includes [a commit that bumps
submodules](https://github.com/tautschnig/verify-rust-std/pull/25/commits/54d1d407a20917a07729abd602d9d0c3f67a335f)).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.