Fix subtree automation for submodules when there were no changes (#396)
While the prior fixes made sure that submodule changes are picked up,
the attempt to commit would fail when there were _no_ changes. See
https://github.com/model-checking/verify-rust-std/actions/runs/15878565075/job/44772389751
for a failing run.

Now we will only attempt to commit when there are changes.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.