Add Flux tool description and CI workflow (#403)
This PR introduces Flux to the list of tools used in the book and adds a
corresponding CI workflow.

The CI workflow runs Flux on a subset of `core` (specifically, files
under `src/ascii`) to verify the absence of array out-of-bounds errors.
It demonstrates how to annotate the code with refinement types to
specify the pre- and post-conditions necessary to prove safety on the
subset that Flux is enabled.

The Flux tool and necessary dependencies are built from source on CI
(pinning specific commits).

Resolves https://github.com/model-checking/verify-rust-std/issues/362

---------

Co-authored-by: Ranjit Jhala <rjhala@ucsd.edu>