Simplify Flux contract for int conversion (#416)
Update the version of Flux and simplify contracts for int conversion as
discussed in
https://github.com/model-checking/verify-rust-std/pull/403#discussion_r2205861328.

Note that this also includes an update in Flux to the toolchain version
matching the one in verify-rust-std, which should fix [this CI
failure](https://github.com/model-checking/verify-rust-std/actions/runs/16326340547/job/46117341963).
Our PR became out of sync between toolchain updates. We apologize for
that!

---------

Co-authored-by: Ranjit Jhala <rjhala@ucsd.edu>