align_to and align_to_mut contract and harnesses (#405)
Towards solving #19 

This PR adds harnesses and contracts for align_to() and align_to_mut().
For the contract for align_to_mut(), we're using a wrapper and writing a
contract for that instead, due to
https://github.com/model-checking/kani/issues/3764.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.