ci: only run update-kani-metrics in model-checking/verify-rust-std (#456)
Fork repos will auto run this job, thus bot PRs are weekly created due
to autoharness and frequent API changes in libstd.

So this commit disables the job for fork repos.



By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.