Bump VeriFast to 25.07 (#453)
The RawVec proof needs 25.07.

Also, a recent upstream change in `linked_list.rs` broke the proof. I
fixed it and upgraded it to 25.07.

To avoid conflicts between the two proof PRs, I'm submitting the
VeriFast bump as a separate PR.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.