VeriFast solution for challenge 19 (RawVec) (#422)
Adds a VeriFast solution for challenge 19 (RawVec).

Note: this proof does not prove semantic well-typedness of the
non-`unsafe` RawVecInner methods listed in the challenge, because they
do not satisfy any meaningful kind of semantic well-typedness; see the
discussion at #283.

Resolves #283

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Felipe R. Monteiro <rms.felipe@gmail.com>